% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{actualizar_archivo}
\alias{actualizar_archivo}
\title{Actualizar un archivo existente en un repositorio GitHub}
\usage{
actualizar_archivo(path_remoto, path_local, sha, repo_owner, repo_name, branch)
}
\arguments{
\item{path_remoto}{Ruta completa del archivo en el repositorio remoto.}

\item{path_local}{Ruta del archivo local.}

\item{sha}{SHA del archivo remoto a actualizar.}

\item{repo_owner}{Usuario o organización dueña del repositorio.}

\item{repo_name}{Nombre del repositorio.}

\item{branch}{Rama del repositorio donde se actualizará el archivo.}
}
\description{
Actualizar un archivo existente en un repositorio GitHub
}
