% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_fuente.R
\name{actualizar_fuente}
\alias{actualizar_fuente}
\title{Actualizar informacion de una fuente}
\usage{
actualizar_fuente(id_fuente, fecha_descarga = NULL, fecha_actualizar = NULL)
}
\arguments{
\item{id_fuente}{integer id numerico que permite seleccionar la fuente segun aparece en el sheet. Para consultar ids usar  `fuentes()`}

\item{fecha_descarga}{string or date Valor de la fecha de descarga. Puede ser clase 'date' como producto de `Sys.Date()` o un string parseable por `as.Date()`}

\item{fecha_actualizar}{string or date Valor de la fecha en que se debe actualizar la descarga. Puede ser clase 'date' como producto de `Sys.Date()` o un string parseable por `as.Date()`}
}
\description{
Actualiza fecha de descarga y acutalizacion de una fuente ya registrada con `agregar_fuente()` en la sheet de fuentes en el drive de Argendata. Hace `drive_upload()` con overwrite = T pisando la version anterior de la fuente.
}
