% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_fuente_raw.R
\name{actualizar_fuente_raw}
\alias{actualizar_fuente_raw}
\title{Actualizar informacion de una fuente raw}
\usage{
actualizar_fuente_raw(id_fuente, fecha_actualizar = NULL)
}
\arguments{
\item{id_fuente}{integer id numerico que permite seleccionar la fuente segun aparece en el sheet. Para consultar ids usar  `fuentes_raw()`}

\item{fecha_actualizar}{string or date Valor de la fecha en que se debe actualizar la descarga. Puede ser clase 'date' como producto de `Sys.Date()` o un string parseable por `as.Date()`}
}
\description{
Actualiza 'fecha_descarga' y 'fecha_actualizar' de una fuente en la sheet de fuentes en el drive de Argendata. Hace `drive_upload()` con overwrite = T pisando la version anterior de la fuente en el drive.
}
\details{
La fecha será actualizada usando `Sys.time()` al momento de su ejecución.
}
