% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_fuente_clean.R
\name{agregar_fuente_clean}
\alias{agregar_fuente_clean}
\title{Agregar nueva fuente clean}
\usage{
agregar_fuente_clean(
  id_fuente_raw = NULL,
  df = NULL,
  path_clean = NULL,
  nombre = NULL,
  script = NULL,
  descripcion = NULL,
  directorio = NULL,
  prompt = TRUE
)
}
\arguments{
\item{id_fuente_raw}{integer id numerico que permite seleccionar la fuente raw segun aparece en el sheet. Para consultar ids usar  \code{fuentes_raw()}}

\item{df}{data.frame Dataframe de la fuente clean a registrar.}

\item{path_clean}{string Nombre del archivo de la fuente tal cual fue guardado.}

\item{nombre}{string Nombre único que identifica a la fuente en su versión 'clean'.}

\item{script}{string  Nombre del archivo del script de descarga de la fuente tal cual se guardó en scripts/limpieza_fuentes/ de argendata-etl}

\item{descripcion}{string Descripcion del dataset}

\item{directorio}{string Ruta al directorio desde el cual cargar el archivo. Si es NULL toma tempdir()}

\item{prompt}{logical Si es TRUE (default) evalua si ya fuentes clean referidas al id_fuente_raw y pide confirmacion antes de continuar.}
}
\description{
Agrega una fuente no registrada previamente: genera una nueva entrada en la sheet de fuentes y hace \code{drive_upload()} con overwrite = F de la fuente.
}
\details{
Más de una fuente clean puede referir a una misma fuente raw. Por ejemplo, si la fuente raw consiste en un excel de multiples hojas, cada hoja debería pasar a ser un csv independiente.
}
