% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argendata_root_dir}
\alias{argendata_root_dir}
\title{Directorio raiz de argendata}
\usage{
argendata_root_dir(version = "v1")
}
\arguments{
\item{version}{string Usar 'v2' para referir al Renv ARGENDATA_DRIVE_V2 (nuevo drive compartido). Sino usar 'v1' (default) para viejo drive argendata}
}
\description{
Directorio raiz de argendata
}
\keyword{internal}
