% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argendataR-package.R
\docType{package}
\name{argendataR-package}
\alias{argendataR}
\alias{argendataR-package}
\title{argendataR: Caja de herramientas para Argendata}
\description{
Paquete de herramientas para el flujo de outputs Argendata.
}
\keyword{internal}
