% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{armador_descripcion}
\alias{armador_descripcion}
\title{Generador de descripcion de columnas}
\usage{
armador_descripcion(metadatos, etiquetas_nuevas = NULL, output_cols)
}
\arguments{
\item{metadatos}{data.frame Subset de variable_nombre y descripcion de metadata.}

\item{etiquetas_nuevas}{data.frame con las columnas variable_nombre y descripcion o lista nombrada con nombre de columna y etiqueta. Ej: list('v_gini' = 'Indice de Gini de los ingresos')}

\item{output_cols}{columnas del dataframe a describir}
}
\value{
list Lista de variables con descripcion
}
\description{
Generador de descripcion de columnas
}
