% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armador_descripcion_alt.R
\name{armador_descripcion_alt}
\alias{armador_descripcion_alt}
\title{armador_descripcion_alt}
\usage{
armador_descripcion_alt(
  metadatos,
  etiquetas_nuevas = data.frame(),
  output_cols
)
}
\arguments{
\item{metadatos}{data.frame sus columnas son variable_nombre y descripcion. Ver \code{argendataR::metadata()}}

\item{etiquetas_nuevas}{data.frame Dataframe con la columna variable_nombre y la descripcion}

\item{output_cols}{string tiene las columnas del dataset que se quiere escribir}
}
\description{
armador_descripcion_alt
}
