% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armar_ruta.R
\name{armar_ruta}
\alias{armar_ruta}
\title{Armar path a un archivo usando normalizePath(glue())}
\usage{
armar_ruta(x)
}
\arguments{
\item{x}{string Cadena de texto con funcionalidades de glue}
}
\value{
string Devuelve un string con la ruta
}
\description{
Armar path a un archivo usando normalizePath(glue())
}
\examples{
armar_ruta("{tempdir()}/archivo.csv")

}
