% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_fuente_clean.R
\name{cargar_fuente_clean}
\alias{cargar_fuente_clean}
\title{(Experimental) Agregar o actualizar una fuente clean}
\usage{
cargar_fuente_clean(
  id_fuente_raw = NULL,
  path_clean = NULL,
  nombre = NULL,
  script = NULL,
  descripcion = NULL,
  directorio = NULL,
  prompt = TRUE,
  actualizar = FALSE
)
}
\arguments{
\item{id_fuente_raw}{integer id numerico que permite seleccionar la fuente raw segun aparece en el sheet. Para consultar ids usar  \code{\link[=fuentes_raw]{fuentes_raw()}}}

\item{path_clean}{string Nombre del archivo de la fuente tal cual fue guardado en el directorio data/_FUENTES/clean/ de argendata-etl}

\item{nombre}{string Nombre único que identifica a la fuente en su versión 'clean'.}

\item{script}{string  Nombre del archivo del script de descarga de la fuente tal cual se guardó en scripts/limpieza_fuentes/ de argendata-etl}

\item{descripcion}{string Descripcion del dataset}

\item{directorio}{string Ruta al directorio desde el cual cargar el archivp}

\item{prompt}{logical Si es TRUE (default) evalua si ya fuentes clean referidas al id_fuente_raw y pide confirmacoion antes de continuar.}

\item{actualizar}{logical Si es TRUE registra la fuente como una nueva mediante \code{\link[=agregar_fuente_clean]{agregar_fuente_clean()}}. Si es FALSE actuliza la  fuente  ya cargada con \code{\link[=actualizar_fuente_raw]{actualizar_fuente_raw()}}.}
}
\description{
Si \code{actualizar == FALSE} (default) Agrega una fuente no registrada previamente: genera una nueva entrada en la sheet de fuentes y hace \code{googledrive::drive_upload()} con overwrite = F de la fuente.
Si \code{actualizar == TRUE} Actualiza 'fecha_descarga' y 'fecha_actualizar' de una fuente en la sheet de fuentes en el drive de Argendata. Hace \code{googledrive::drive_upload()} con overwrite = T pisando la version anterior de la fuente en el drive.
}
