% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_fuente_raw.R
\name{cargar_fuente_raw}
\alias{cargar_fuente_raw}
\title{(Experimental) Agregar o actualizar una fuente raw}
\usage{
cargar_fuente_raw(
  url = NULL,
  nombre = NULL,
  institucion = NULL,
  actualizable = NULL,
  fecha_descarga = NULL,
  fecha_actualizar = NULL,
  path_raw = NULL,
  script = NULL,
  api = FALSE,
  actualizar = FALSE,
  directorio = NULL
)
}
\arguments{
\item{url}{string Link directo a la fuente si existiera o link a la página web más inmediata a la  fuente.}

\item{nombre}{string Nombre único que identifica a la fuente}

\item{institucion}{string Nombre oficial de la institucion}

\item{actualizable}{logical TRUE o FALSE  sobre si la fuente será actualizada y debe volver a ser descargada en nueva versión en el futuro.}

\item{fecha_descarga}{date o string o null Fecha de descarga como valor de clase 'date', o 'string' parseable por \code{as.Date()}. Si es null toma la fecha de \code{Sys.Date()}}

\item{fecha_actualizar}{date o string o null Fecha de descarga como valor de clase 'date', o 'string' parseable por \code{as.Date()}. Si es null toma fecha actual más 6 meses}

\item{path_raw}{string Nombre del archivo de la fuente tal cual fue descargado en el directorio data/_FUENTES/raw/ de argendata-etl}

\item{script}{string  Nombre del archivo del script de descarga de la fuente tal cual se guardó en scripts/descarga_fuentes/ de argendata-etl}

\item{api}{logical TRUE o FALSE indicando si la fuente es una api o no.}

\item{actualizar}{logical Si es TRUE registra la fuente como una nueva mediante \code{\link[=agregar_fuente_raw]{agregar_fuente_raw()}}. Si es FALSE actuliza la  fuente  ya cargada con \code{\link[=actualizar_fuente_raw]{actualizar_fuente_raw()}}}

\item{directorio}{string Ruta al directorio desde el cual cargar el archivp}
}
\description{
Si \code{actualizar == FALSE} (default) Agrega una fuente no registrada previamente: genera una nueva entrada en la sheet de fuentes y hace \code{googledrive::drive_upload()} con overwrite = F de la fuente.
Si \code{actualizar == TRUE} Actualiza 'fecha_descarga' y 'fecha_actualizar' de una fuente en la sheet de fuentes en el drive de Argendata. Hace \code{googledrive::drive_upload()} con overwrite = T pisando la version anterior de la fuente en el drive.
}
