% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_outputs.R
\name{check_datatypes}
\alias{check_datatypes}
\title{Control de clases de outputs de argendata}
\usage{
check_datatypes(df, df_anterior)
}
\arguments{
\item{df}{dataframe output nuevo}

\item{df_anterior}{dataframe output previo}
}
\value{
dataframe de control de clases
}
\description{
Control de clases de outputs de argendata
}
