% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_iso3.R
\name{check_iso3}
\alias{check_iso3}
\title{Verifica códigos ISO3}
\usage{
check_iso3(x)
}
\arguments{
\item{x}{Un vector de caracteres que contiene los códigos ISO3 a verificar.}
}
\value{
Devuelve un valor lógico TRUE si todos los códigos están validados. Si no, devuelve un vector lógico indicando cuáles códigos no son válidos.
}
\description{
Esta función verifica si todos los elementos de un vector están presentes en un conjunto de códigos ISO3 validados.
}
\examples{
# Ejemplo de uso
check_iso3(c("ARG", "BRA", "USA"))

}
