% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_to}
\alias{coerce_to}
\title{Coercion de un objeto a la clase de otro objeto}
\usage{
coerce_to(x, y)
}
\arguments{
\item{x}{objeto a coercionar}

\item{y}{objeto de clase objetivo}
}
\value{
objeto x coercionado a clase y
}
\description{
Coercion de un objeto a la clase de otro objeto
}
