% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colectar_fuentes.R
\name{colectar_fuentes}
\alias{colectar_fuentes}
\title{colectar_fuentes}
\usage{
colectar_fuentes(pattern = "^fuente.*")
}
\arguments{
\item{pattern}{string regex para capturar los nombres de variables en el entorno global}
}
\description{
colectar_fuentes
}
