% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_outputs.R
\name{control_valores_nonnum}
\alias{control_valores_nonnum}
\title{Controles para variables no numericas}
\usage{
control_valores_nonnum(root_name, pk, df)
}
\arguments{
\item{root_name}{raiz del nombre de la columna. Ej. para comparar "valores.x" y "valores.y" se usa "valores"}

\item{pk}{primary keys de del dataframe}

\item{df}{dataframe}
}
\value{
lista con los resultados de los controles
}
\description{
Compara dos columnas de un dataframe, en general que es producto de un join.
Las columnas a comparar deben estar nombradas al estilo "colA.x" y "colA.y"
}
\details{
Los valores devueltos son:
\itemize{
\item nuevos_na: cantidad de nuevos na en la col.y respecto a la col.x
\item tasa_mismatches: proporcion de mismathces (filas que no coinciden entre col.x y col.y) sobre el total de filas
\item metricas_filas =  dataframe con pk, col.x, col.y, coinciden_valores
}
}
