% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_fuente.R
\name{descargar_fuente}
\alias{descargar_fuente}
\title{Descarga de fuentes de Argendata desde el drive}
\usage{
descargar_fuente(codigo, dir = NULL)
}
\arguments{
\item{codigo}{string Codigo de la fuente tal cual aparece en la consulta a}

\item{dir}{string directorio donde se descarga la fuente. Si es NULL toma \code{getwd()}}
}
\value{
Descarga el archivo correspondiente al codigo de fuente desde el drive de Argendata
}
\description{
Descarga de fuentes de Argendata desde el drive
}
