% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_fuente_raw.R
\name{descargar_fuente_raw}
\alias{descargar_fuente_raw}
\title{Descarga fuente version raw}
\usage{
descargar_fuente_raw(id_fuente, dir = NULL, limpiar_cache = F)
}
\arguments{
\item{id_fuente}{numeric id numerico de la fuente tal cual aparece en la sheet de fuentes. Ver \code{fuentes_raw()}}

\item{dir}{string directorio donde se descarga la fuente}

\item{limpiar_cache}{Logical Si es TRUE borra el cache de consultas al drive de argendata. Si es FALSE reutiliza el cache existente.}
}
\value{
file la fuente seleccionada de Fuentes/raw descargada en el directorio especificado
}
\description{
Descarga fuente version raw
}
