% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_output.R
\name{descargar_output}
\alias{descargar_output}
\title{Descargar output desde el drive}
\usage{
descargar_output(
  nombre,
  subtopico,
  entrega_subtopico = NULL,
  branch = "main",
  ...
)
}
\arguments{
\item{nombre}{nombre del archivo a descargar. La función usa grepl para identificar el archivo, si el nombre coincide con más de un archivo devuelve error.}

\item{subtopico}{codigo de 6 letras del subtopico en mayusculas.}

\item{entrega_subtopico}{nombre exacto de la carpeta de entrega donde buscar el output}

\item{branch}{branch de github a donde apuntar. Default es "main"}

\item{...}{parametros adicionales pasados a read_delim}
}
\description{
La funcion descarga el archivo si está disponible desde GH y sino desde la carpeta del drive de argendata
en la carpeta temporal de la sesión y hace read_delim desde allí.
}
