% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansor_xvar.R
\name{expansor_xvar}
\alias{expansor_xvar}
\title{Title}
\usage{
expansor_xvar(x, var)
}
\arguments{
\item{var}{}
}
\description{
para empalmar una serie usando variaciones de otra serie
dada una columna x con los valores de la serie a expandir: Xi a Xn de 1 a N
y dada una columna var donde VARi = Xi/Xi-1 como las proporciones entre un valor de x y su antecedente
si Xi es NA lo calcula y reemplaza como Xi-1*VARi, si no es NA lo deja tal cual
}
