% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuentes.R
\name{fuentes}
\alias{fuentes}
\title{Fuentes}
\usage{
fuentes()
}
\arguments{
\item{limpiar_cache}{Logical Si es TRUE borra el cache de consultas al drive de argendata. Si es FALSE reutiliza el cache existente.}
}
\value{
tibble con la sheet de fuentes raw+clean de googledrive Argendata
}
\description{
Fuentes
}
