% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fuentes_clean_dir}
\alias{fuentes_clean_dir}
\title{lista de entradas dentro de Fuentes/clean}
\usage{
fuentes_clean_dir()
}
\value{
tibble con el directorio de fuentes clean
}
\description{
lista de entradas dentro de Fuentes/clean
}
\keyword{internal}
