% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fuentes_clean_sheet_id}
\alias{fuentes_clean_sheet_id}
\title{fuentes_raw_sheet_id}
\usage{
fuentes_clean_sheet_id()
}
\value{
id de sheet de fuentes
}
\description{
fuentes_raw_sheet_id
}
\keyword{internal}
