% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuentes_raw.R
\name{fuentes_raw}
\alias{fuentes_raw}
\title{Fuentes raw}
\usage{
fuentes_raw(limpiar_cache = T)
}
\arguments{
\item{limpiar_cache}{Logical Si es TRUE borra el cache de consultas al drive de argendata. Si es FALSE reutiliza el cache existente.}
}
\value{
tibble con la sheet de fuentes raw de googledrive Argendata
}
\description{
Fuentes raw
}
