% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fuentes_raw_dir}
\alias{fuentes_raw_dir}
\title{lista de entradas dentro de Fuentes/raw}
\usage{
fuentes_raw_dir()
}
\value{
tibble con el directorio de fuentes raw
}
\description{
lista de entradas dentro de Fuentes/raw
}
\keyword{internal}
