% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_path.R
\name{get_fuente_path}
\alias{get_fuente_path}
\title{Consulta el path de fuentes}
\usage{
get_fuente_path(codigo = NULL)
}
\arguments{
\item{codigo}{string Codigo de fuente cuyo path se busca. Ej: \code{"R37C1"} o \code{"R37C0"}}
}
\value{
string Path de la fuente
}
\description{
Consulta el path de fuentes
}
