% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids_graficos.R
\name{get_ids_graficos}
\alias{get_ids_graficos}
\title{Obtener IDs de gráficos de ArgenData}
\usage{
get_ids_graficos(url)
}
\arguments{
\item{url}{Un vector o lista de caracteres con las URLs de donde se extraerán los IDs de los gráficos}
}
\value{
Un vector de caracteres con los IDs de los gráficos, ordenados alfabéticamente y sin duplicados
}
\description{
Extrae los identificadores de gráficos disponibles en una o varias URLs de ArgenData
}
\examples{
\dontrun{
ids <- get_ids_graficos("https://argendata.fund.ar/notas/que-es-la-intensidad-de-carbono/")
}
}
