% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapping.R
\name{get_mapping}
\alias{get_mapping}
\title{get_mapping}
\usage{
get_mapping(subtopico)
}
\arguments{
\item{subtopico}{string Nombre de 6 letras del subtopico}
}
\value{
dataframe Devuelve un dataframe con el mapping del subtopico
}
\description{
Devuelve un dataframe con el mapping del subtopico publicado en main del repo transformers
}
