% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomenclador_paises.R
\name{get_nomenclador_geografico}
\alias{get_nomenclador_geografico}
\title{Devuelve el nomenclador geografico para argendata}
\usage{
get_nomenclador_geografico()
}
\value{
tibble con la sheet de consolidado_fundar del drive de Argendata
}
\description{
Devuelve el nomenclador geografico para argendata
}
