% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomenclador_geografico_front.R
\name{get_nomenclador_geografico_front}
\alias{get_nomenclador_geografico_front}
\title{Devuelve el nomenclador geografico para argendata}
\usage{
get_nomenclador_geografico_front()
}
\value{
tibble con la sheet de consolidado_fundar del drive de Argendata
}
\description{
Devuelve el nomenclador geografico para argendata
}
