% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomenclador_paises.R
\name{get_nomenclador_paises}
\alias{get_nomenclador_paises}
\title{get_nomenclador_paises}
\usage{
get_nomenclador_paises()
}
\value{
tibble con la sheet de consolidado_fundar del drive de Argendata
}
\description{
get_nomenclador_paises
}
