% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_output.R
\name{get_output_repo}
\alias{get_output_repo}
\title{Descargar output desde el repo de data de argendata}
\usage{
get_output_repo(path, subtopico, branch = "main")
}
\arguments{
\item{path}{path del archivo a descargar. Ej.: '01_evolucion_CO2_historico.json' o '01_evolucion_CO2_historico.csv'}

\item{subtopico}{codigo de 6 letras del subtopico en mayusculas. Ej.: "CAMCLI"}

\item{branch}{branch de github a donde apuntar. Default es "main"}
}
\description{
La funcion descarga el archivo si está disponible desde GH y sino desde la carpeta del drive de argendata
en la carpeta temporal de la sesión y hace read_delim desde allí.
}
