% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_raw.R
\name{get_raw_path}
\alias{get_raw_path}
\title{Devuelve el full path de un codigo de fuente raw}
\usage{
get_raw_path(codigo)
}
\arguments{
\item{codigo}{string Codigo de la fuente raw para la cual se busca el path de referencia. Ej: \code{"R\\\\d{1,2}C0"}}
}
\description{
Devuelve el full path de un codigo de fuente raw
}
