% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_path.R
\name{get_temp_path}
\alias{get_temp_path}
\title{Consulta los full path de archivos en el directorio temporal}
\usage{
get_temp_path(pattern = NULL)
}
\arguments{
\item{pattern}{string Patron de texto del nombre de archivo a recuperar. Ej: \code{"R37C1"} o \code{"R\\\\d{1,2}C0"}}
}
\description{
Consulta los full path de archivos en el directorio temporal
}
