% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids_graficos.R
\name{inner_get_ids_graficos}
\alias{inner_get_ids_graficos}
\title{Función interna para obtener IDs de gráficos}
\usage{
inner_get_ids_graficos(url)
}
\arguments{
\item{url}{Una cadena de caracteres con la URL de donde se extraerán los IDs de los gráficos}
}
\value{
Un vector de caracteres con los IDs de los gráficos encontrados
}
\description{
Función auxiliar que extrae los identificadores de gráficos de una única URL de ArgenData
}
\keyword{internal}
