% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_temps.R
\name{limpiar_temps}
\alias{limpiar_temps}
\title{Borra archivos temporales usados por el paquete argendataR}
\usage{
limpiar_temps()
}
\description{
Borra archivos temporales usados por el paquete argendataR
}
