% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{mandar_data}
\alias{mandar_data}
\title{Subir o actualizar un archivo en repositorio data de argendatafundar}
\usage{
mandar_data(nombre_archivo, branch = "dev")
}
\arguments{
\item{nombre_archivo}{Nombre del archivo a subir o actualizar.}

\item{branch}{Rama del repositorio donde se realizará la operación Por defecto, 'main'.}
}
\description{
Subir o actualizar un archivo en repositorio data de argendatafundar
}
