% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Consulta metadata de subtopicos}
\usage{
metadata(subtopico = NULL)
}
\arguments{
\item{subtopico}{string Texto con el codigo de 6 letras del subtopico o patron de regex}
}
\value{
dataframe con la metadata
}
\description{
Consulta metadata de subtopicos
}
