% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_viejo}
\alias{metadata_viejo}
\title{Consulta metadata de subtopicos viejos}
\usage{
metadata_viejo(subtopico = NULL, version = NULL, skip = 0)
}
\arguments{
\item{subtopico}{string Texto con el codigo de 6 letras del subtopico o patron de regex}

\item{version}{(deprecated) Solo a fines de compatibilidad}

\item{skip}{lineas a saltear en la lectura del sheet de metada}
}
\value{
dataframe con la metadata
}
\description{
Consulta metadata de subtopicos viejos
}
