% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_path}
\alias{normalize_path}
\title{normalize path con supress warning}
\usage{
normalize_path(path)
}
\value{
normalized path
}
\description{
normalize path con supress warning
}
\keyword{internal}
