% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{obtener_fila_max_tiempo}
\alias{obtener_fila_max_tiempo}
\title{Obtener la fila con los valores más altos de las columnas de tiempo}
\usage{
obtener_fila_max_tiempo(data, columnas_tiempo)
}
\arguments{
\item{data}{Un dataframe que contiene las columnas de tiempo.}

\item{columnas_tiempo}{Un vector de caracteres con los nombres de las columnas
de tiempo que se utilizarán para ordenar el dataframe.}
}
\value{
Una lista con los nombres y valores de las columnas de tiempo más altos.
}
\description{
Esta función ordena un dataframe según las columnas de tiempo especificadas
y devuelve una lista con los valores más altos de estas columnas.
}
\examples{
# Ejemplo de uso
data <- data.frame(anio = c(2020, 2021, 2022), mes = c(5, 6, 7))
columnas_tiempo <- c("anio", "mes")
obtener_fila_max_tiempo(data, columnas_tiempo)

}
