% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_scraper.R
\name{owid_scraper}
\alias{owid_scraper}
\title{Scraper de datos de OWID}
\usage{
owid_scraper(url)
}
\arguments{
\item{url}{string con url al grafico de OWID del que se quieren tomar los datos}
}
\value{
dataframe Devuelve un dataframe con los datos y algunos metadatos
}
\description{
Scraper de datos de OWID
}
