% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fuente_clean.R
\name{read_fuente_clean}
\alias{read_fuente_clean}
\title{Lee una fuente de fuentes clean de Argendata}
\usage{
read_fuente_clean(id_fuente)
}
\arguments{
\item{id_fuente}{string con el codigo o id numerico con la fuente tal cual aparece en \code{fuentes_clean()}}
}
\value{
dataframe correspondiente al codigo de fuente clean de Argendata
}
\description{
Lee una fuente de fuentes clean de Argendata
}
