% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet_srv.R
\name{read_parquet_srv}
\alias{read_parquet_srv}
\title{Lee una fuente de fuentes de Argendata desde el drive}
\usage{
read_parquet_srv(id_fuente)
}
\arguments{
\item{id_fuente}{string con el codigo o id numerico con la fuente tal cual aparece en \code{fuentes_clean()}}
}
\value{
Descarga el archivo correspondiente al codigo de fuente desde el drive de Argendata
}
\description{
Lee una fuente de fuentes de Argendata desde el drive
}
