% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_fuente_clean.R
\name{script_fuente_clean}
\alias{script_fuente_clean}
\title{Crea un .R con el esquema basico de script para outputs}
\usage{
script_fuente_clean(path = NULL, .navigate = F)
}
\arguments{
\item{path}{string Ruta y nombre del archivo a crear}

\item{.navigate}{logical Si es TRUE abre el editor del archivo creado. FALSE (default).}
}
\value{
Crea un archivo .R con el esquema basico para un script de subtopico
}
\description{
Crea un .R con el esquema basico de script para outputs
}
