% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_fuentes.R
\name{script_fuentes}
\alias{script_fuentes}
\title{Crea un .R con el esquema basico de script para la descarga de fuentes desde el drive de Argendata}
\usage{
script_fuentes(path = NULL, .navigate = F)
}
\arguments{
\item{path}{string Ruta y nombre del archivo a crear}

\item{.navigate}{logical Si es TRUE abre el editor del archivo creado. FALSE (default).}
}
\value{
Crea un archivo .R con el esquema basico para un script de subtopico
}
\description{
Crea un .R con el esquema basico de script para la descarga de fuentes desde el drive de Argendata
}
