% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_subtopico.R
\name{script_subtopico}
\alias{script_subtopico}
\title{Crea un .R con el esquema basico de script para outputs}
\usage{
script_subtopico(path = NULL)
}
\arguments{
\item{path}{string Ruta y nombre del archivo a crear}
}
\value{
Crea un archivo .R con el esquema basico para un script de subtopico
}
\description{
Crea un .R con el esquema basico de script para outputs
}
