% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{subir_o_actualizar}
\alias{subir_o_actualizar}
\title{Subir o actualizar un archivo en un repositorio GitHub}
\usage{
subir_o_actualizar(nombre_archivo, repo_owner, repo_name, branch = "main")
}
\arguments{
\item{nombre_archivo}{Nombre del archivo a subir o actualizar.}

\item{repo_owner}{Usuario o organizaci칩n due침a del repositorio.}

\item{repo_name}{Nombre del repositorio.}

\item{branch}{Rama del repositorio donde se realizar치 la operaci칩n Por defecto, 'main'.}
}
\description{
Subir o actualizar un archivo en un repositorio GitHub
}
