% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_fundar.R
\name{write_csv_fundar}
\alias{write_csv_fundar}
\title{write_csv estilo Fundar}
\usage{
write_csv_fundar(x, file)
}
\arguments{
\item{x}{un dataframe a escribir como csv}

\item{file}{nombre del archivo a escribir}
}
\value{
un archivo
}
\description{
write_csv estilo Fundar
}
