% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_fuente_clean.R
\name{actualizar_fuente_clean}
\alias{actualizar_fuente_clean}
\title{Actualizar informacion de una fuente clean}
\usage{
actualizar_fuente_clean(
  id_fuente_clean,
  nombre = NULL,
  script = NULL,
  path_clean = NULL,
  descripcion = NULL,
  directorio = NULL,
  prompt = T
)
}
\arguments{
\item{id_fuente_clean}{integer id numerico que permite seleccionar la fuente segun aparece en el sheet. Para consultar ids usar  \code{fuentes_clean()}}

\item{nombre}{string Nombre único que identifica a la fuente en su versión 'clean'.}

\item{script}{string  Nombre del archivo del script de descarga de la fuente tal cual se guardó en scripts/limpieza_fuentes/ de argendata-etl}

\item{path_clean}{string Nombre del archivo de la fuente tal cual fue guardado.}

\item{descripcion}{string Descripcion del dataset}

\item{directorio}{string Ruta al directorio desde el cual cargar el archivo. Si es NULL toma tempdir()}

\item{prompt}{logical Si es TRUE (default) pide confirmacion de los cambios.}
}
\description{
Hace \code{drive_upload()} con overwrite = T pisando la version anterior de la fuente en el drive y actualiza 'timestamp' de una fuente en la sheet de fuentes clean en el drive de Argendata.
}
\details{
La funcion toma el dato de path_clean declarado en \code{fuentes_clean()} para el id seleccionado y espera que exista el archivo correpondiente al path_clean. Ese archivo ser subirá al drive pisando el archivo preexistente.
La fecha será actualizada usando \code{Sys.time()} al momento de su ejecución.
}
