% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_fuente_clean.R
\name{actualizar_fuente_clean}
\alias{actualizar_fuente_clean}
\title{Actualizar informacion de una fuente clean}
\usage{
actualizar_fuente_clean(id_fuente_clean, fecha = NULL)
}
\arguments{
\item{id_fuente_clean}{integer id numerico que permite seleccionar la fuente segun aparece en el sheet. Para consultar ids usar  `fuentes_raw()`}

\item{fecha}{string or date Valor de la fecha de descarga. Puede ser clase 'date' como producto de `Sys.Date()` o un string parseable por `as.Date()`}
}
\description{
Actualiza 'fecha' de una fuente en la sheet de fuentes clean en el drive de Argendata y hace `drive_upload()` con overwrite = T pisando la version anterior de la fuente en el drive.
}
