% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_fuente_clean.R
\name{actualizar_fuente_clean}
\alias{actualizar_fuente_clean}
\title{Actualizar informacion de una fuente clean}
\usage{
actualizar_fuente_clean(
  id_fuente_clean,
  df = NULL,
  nombre = NULL,
  script = NULL,
  path_clean = NULL,
  descripcion = NULL,
  directorio = NULL,
  prompt = T
)
}
\arguments{
\item{id_fuente_clean}{integer id numerico  o codigo string que permite seleccionar la fuente segun aparece en el sheet. Para consultar ids usar  \code{fuentes_clean()}}

\item{df}{data.frame Datafrane de la fuente clean a registrar.}

\item{nombre}{string Nombre único que identifica a la fuente en su versión 'clean'.}

\item{script}{string  Nombre del archivo del script de descarga de la fuente tal cual se guardó en scripts/limpieza_fuentes/ de argendata-etl}

\item{path_clean}{string Nombre del archivo de la fuente. Debe ser extension .parquet.}

\item{descripcion}{string Descripcion del dataset}

\item{directorio}{string Ruta al directorio desde el cual cargar el archivo. Si es NULL toma tempdir()}

\item{prompt}{logical Si es TRUE (default) pide confirmacion de los cambios.}
}
\description{
Actualiza el registro en la tabla de fuentes clean y guarda una nueva copia de la fuente clean el directorio del servidor.
}
\details{
La funcion toma el dato de path_clean declarado en \code{fuentes_clean()} para el id seleccionado y espera que exista el archivo correpondiente al path_clean. Ese archivo ser subirá al drive pisando el archivo preexistente.
La fecha será actualizada usando \code{Sys.time()} al momento de su ejecución.
}
