% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argendata_root_dir}
\alias{argendata_root_dir}
\title{Directorio raiz de argendata}
\usage{
argendata_root_dir(version)
}
\arguments{
\item{version}{(deprecated) Queda solo a fines de compatibilidad}
}
\description{
Directorio raiz de argendata
}
\keyword{internal}
