% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_outputs.R
\name{check_cols}
\alias{check_cols}
\title{Control de nombres de outputs de argendata}
\usage{
check_cols(df, df_anterior)
}
\arguments{
\item{df}{dataframe nuevo de output}

\item{df_anterior}{dataframe anterior para la comparacion}
}
\value{
lista de resultados
}
\description{
Control de nombres de outputs de argendata
}
