% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_archivos.R
\name{comparar_archivos}
\alias{comparar_archivos}
\title{Comparar dos archivos}
\usage{
comparar_archivos(x, y)
}
\arguments{
\item{x}{string path a un archivo}

\item{y}{string path a un archivo}
}
\value{
Devuelve una lista con los resultados de los chequeos realizados
}
\description{
Compara dos archivos y devuelve una lista de controles
}
