% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_outputs.R
\name{comparar_outputs}
\alias{comparar_outputs}
\title{Comparar un data.frame con un output cargado en el drive de Argendata como csv}
\usage{
comparar_outputs(
  df,
  nombre,
  subtopico,
  entrega_subtopico = "primera_entrega",
  pk = NULL,
  drop_output_drive = T
)
}
\arguments{
\item{df}{data.frame dataframe a comparar con el output cargado en el drive}

\item{nombre}{string nombre del output en el drive}

\item{subtopico}{string subtopico al que pertenece el output}

\item{entrega_subtopico}{string nombre de la carpeta de entrega donde buscar el output}

\item{pk}{string Vector con los nombres de columnas que son primary key del dataframe}

\item{drop_output_drive}{logical Si es TRUE (defatult) el resultado incluye el output cargado en el drive usado para comparar.}
}
\value{
Devuelve una lista con los resultados de los chequeos realizados
}
\description{
Comparar un data.frame con un output cargado en el drive de Argendata como csv
}
