% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_outputs.R
\name{control_valores_num}
\alias{control_valores_num}
\title{Controles para variables numericas}
\usage{
control_valores_num(x, pk, df)
}
\arguments{
\item{x}{raiz del nombre de la columna. Ej. para comparar "valores.x" y "valores.y" se usa "valores"}

\item{pk}{primary keys de del dataframe}

\item{df}{dataframe}
}
\value{
lista con los resultados de los controles
}
\description{
Compara dos columnas de un dataframe, en general que es producto de un join.
Las columnas a comparar deben estar nombradas al estilo "colA.x" y "colA.y"
}
\details{
Los valores devueltos son:
\itemize{
\item nuevos_na: cantidad de nuevos na en la col.y respecto a la col.x
\item mean_variaciones_rel: promedio de las variaciones relativas (col.y/col.x)-1
\item ks_test: p valor del Kolmogorov-Smirnov Tests entre col.x y col.y
\item mw_test: p valor del Mann-Whitney test entre col.x y col.y
qqplot_var  = qqplot(col.x, col.y)
\item metricas_filas =  dataframe con pk, col.x, col.y, col_var
}
}
