% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_fuente_raw.R
\name{descargar_fuente_raw}
\alias{descargar_fuente_raw}
\title{Descarga fuente version raw}
\usage{
descargar_fuente_raw(id_fuente, nombre, dir)
}
\arguments{
\item{id_fuente}{id numerico de la fuente tal cual aparece en la sheet de fuentes. Ver `fuentes_raw()`}

\item{nombre}{nombre de la fuente tal cual aparece en la sheet de fuentes. Ver `fuentes_raw()`}

\item{dir}{directorio donde se descarga la fuente}
}
\value{
file la fuente seleccionada de Fuentes/raw descargada en el directorio especificado
}
\description{
Descarga fuente version raw
}
