% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansor_xvar.R
\name{expansor_xvar}
\alias{expansor_xvar}
\title{Expandir una serie usando variaciones de otra serie}
\usage{
expansor_xvar(x, var)
}
\arguments{
\item{x}{variable con serie a expandir}

\item{var}{variable con variaciones a usar para expandir la  serie}
}
\value{
vector con la serie expandida
}
\description{
Toma un vector x con valores definidos de x1 a xn y con valores faltantes de xn+1 a xm y calcula los valores faltantes usando variaciones de otro vector var1 a varm como: xn+1 = xn*varn+1.
}
