% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_subtopico.R
\name{init_subtopico}
\alias{init_subtopico}
\title{Crear sripts y md del subtopico}
\usage{
init_subtopico(subtopico)
}
\arguments{
\item{subtopico}{string codigo de 6 letras del subtopico}
}
\description{
Crear sripts y md del subtopico
}
