% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{mandar_data}
\alias{mandar_data}
\title{Subir o actualizar un archivo en repositorio data de argendatafundar}
\usage{
mandar_data(filename, subtopico, branch = "dev")
}
\arguments{
\item{filename}{Nombre del archivo a buscar en tmpdir.}

\item{subtopico}{Ruta en el repositorio remoto donde se guardará el archivo.}

\item{branch}{Rama del repositorio donde se realizará la operación Por defecto, 'dev'.}
}
\description{
Subir o actualizar un archivo en repositorio data de argendatafundar
}
