% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_path}
\alias{normalize_path}
\title{Directorio fuentes de argendata}
\usage{
normalize_path(path)
}
\value{
id de sheet de fuentes
fuentes_raw_sheet_id

id de sheet de fuentes
lista de entradas dentro de Fuentes/raw

tibble con el directorio de fuentes raw

tibble con el directorio de fuentes clean

normalized path
}
\description{
Directorio fuentes de argendata
}
\keyword{Fuentes/clean}
\keyword{con}
\keyword{de}
\keyword{dentro}
\keyword{entradas}
\keyword{fuentes_raw_sheet_id}
\keyword{internal}
\keyword{lista}
\keyword{normalize}
\keyword{path}
\keyword{supress}
\keyword{warning}
