% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{subir_archivo}
\alias{subir_archivo}
\title{Subir un nuevo archivo a un repositorio GitHub}
\usage{
subir_archivo(path_remoto, path_local, repo_owner, repo_name, branch)
}
\arguments{
\item{path_remoto}{Ruta completa del archivo en el repositorio remoto.}

\item{path_local}{Ruta del archivo local.}

\item{repo_owner}{Usuario o organizaci\u00f3n due\u00f1a del repositorio.}

\item{repo_name}{Nombre del repositorio.}

\item{branch}{Rama del repositorio donde se subir\u00e1 el archivo.}
}
\description{
Subir un nuevo archivo a un repositorio GitHub
}
