% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{subir_o_actualizar}
\alias{subir_o_actualizar}
\title{Subir o actualizar un archivo en un repositorio GitHub}
\usage{
subir_o_actualizar(
  path_local,
  path_remoto,
  repo_owner,
  repo_name,
  branch = "main"
)
}
\arguments{
\item{path_local}{Ruta completa del archivo local. (Por ejemplo '/tmp/data.json')}

\item{path_remoto}{Ruta en donde se subirá o actualizará el archivo en el repositorio remoto
(Por ejemplo 'CRECIM/data.json')}

\item{repo_owner}{Usuario o organización dueña del repositorio.}

\item{repo_name}{Nombre del repositorio.}

\item{branch}{Rama del repositorio donde se realizará la operación Por defecto, 'main'.}
}
\description{
Subir o actualizar un archivo en un repositorio GitHub
}
