% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_subtopico.R
\name{subtopico_outputs}
\alias{subtopico_outputs}
\title{Consulta lista de outputs del subtopico/entrega en drive}
\usage{
subtopico_outputs(subtopico_nombre, entrega_subtopico)
}
\arguments{
\item{subtopico_nombre}{codigo 6 letras del subtopico tal cual su nombre en el drive de Argendata}

\item{entrega_subtopico}{nombre de la entrega a buscar tal cual su nombre en el drive de Argendata}
}
\value{
dataframe con los outputs del subtopico para la entrega elegida
}
\description{
Consulta lista de outputs del subtopico/entrega en drive
}
